#8 Add リストから削除可能にする（Ajaxはまだ）
