+++ app/views/artists/_added_artist.html.erb
@@ -1,4 +1,4 @@
-<div class="my-1 px-1 w-1/3 md:w-1/4 lg:my-4 lg:px-1 lg:w-1/5 mb-10 mt-2 relative">
+<div class="my-1 px-1 w-1/3 md:w-1/4 lg:my-4 lg:px-1 lg:w-1/5 mb-10 mt-2 relative" id="my-artist-<%= artist["id"] %>">
     <%= image_tag(artist["image_url"], class: "block w-full" ) %>
     <p><%= artist["name"] %></p>
     <%= form_with(scope: :artist_info, url: favorite_artist_path(artist["id"]), remote: true, method: :delete) do |f| %>
